requirements.txt
ipykernel 
openai
langchain
python-dotenv
streamlit
langchain_community
langchain-openai